requirements.txt
# Core dependencies
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.1.0
grpcio>=1.54.2
grpcio-tools>=1.54.2
grpcio-health-checking>=1.54.2
grpcio-reflection>=1.54.2
protobuf>=4.23.2

# Scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Bio-informatics
biopython>=1.79
rdkit>=2022.9.3

# Cloud and deployment
boto3>=1.26.0
botocore>=1.29.0
awscli>=1.27.0
kubernetes>=26.1.0
redis>=4.5.0

# Utilities
click>=8.0.0
pyyaml>=6.0.0
tqdm>=4.62.0
requests>=2.26.0
einops>=0.7.0
fairscale>=0.4.13
mashumaro>=3.11

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.0.0
black>=22.0.0
isort>=5.0.0
flake8>=3.9.0
mypy>=0.910
